feat: 포인트 기능 추가

+++ src/main/java/com/example/gotsaeng_back/domain/auth/dto/PointDto.java
@@ -11,6 +11,6 @@ import java.time.LocalDate;
 @Setter
 @AllArgsConstructor
 public class PointDto {
-    private double value;
+    private long value;
     private LocalDate getDate;
 }
