+++ src/main/java/com/example/gotsaeng_back/domain/auth/entity/Point.java
@@ -27,7 +27,7 @@ public class Point {
     private LocalDate getDate;
 
     @Column(name = "value", nullable = false)
-    private double value;
+    private long value;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
